Class for 'git ls', and memoize
